Introduce MouseInput.h
