+++ src/openrct2-ui/WindowManager.cpp
@@ -14,6 +14,7 @@
 #include "windows/Window.h"
 
 #include <openrct2-ui/input/InputManager.h>
+#include <openrct2-ui/input/MouseInput.h>
 #include <openrct2-ui/input/ShortcutManager.h>
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Input.h>
